# docs/swagger.yaml
basePath: /
definitions:
...
  handler.UpdateStatusInput:
    properties:
      is_active:
        example: true
        type: boolean
    type: object
  handler.UserResponse:
    properties:
      id:
        example: 1
        type: integer
      identifier:
        example: GR001
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Ahmad Guru
        type: string
      role:
        example: guru
        type: string
    type: object
  helpers.SwaggerError400Response:
    properties:
      message:
        example: Input tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError401Response:
    properties:
      message:
        example: Token tidak ditemukan atau tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError403Response:
    properties:
      code:
        example: ERR_PERMISSION_DENIED
        type: string
      message:
        example: 'Akses ditolak: Anda tidak memiliki akses untuk fitur ini'
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError404Response:
    properties:
      message:
        example: Data tidak ditemukan
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError500Response:
    properties:
      message:
        example: Terjadi kesalahan internal pada server
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerSuccessResponse:
    properties:
      data: {}
      message:
        example: Berhasil memproses data
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.User:
    properties:
      alamat:
        example: Jl. Diponegoro No. 25, Salatiga
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      email:
        example: ahmad.guru@sekolah.com
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_first_login:
        example: true
        type: boolean
      jenis_kelamin:
        example: Laki-laki
        type: string
      last_login_at:
        type: string
      nama:
        example: Ahmad Guru
        type: string
      nip:
        description: Ignored by GORM, populated for unified API login response
        example: GR001
        type: string
      nis:
        example: "12345678"
        type: string
      no_hp:
        example: "081234567890"
        type: string
      role:
        example: guru
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: praktikum@example.com
    name: Praktikum Pemrograman III
  description: Dokumentasi API backend Sistem Informasi Sekolah menggunakan Golang
    Fiber, GORM, PostgreSQL, dan JWT.
  title: API Sistem Informasi Sekolah
  version: "1.0"
paths:
...
  /api/users/:
    get:
      consumes:
      - application/json
      description: Mengambil daftar seluruh user account (Admin, Guru, Siswa) beserta
        nama asli dan identifier (NIP/NIS/Email). Hanya dapat diakses oleh Admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.SwaggerSuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil semua data user account
      tags:
      - User Management
  /api/users/{id}/reset-password:
    put:
      consumes:
      - application/json
      description: Mereset password user berdasarkan ID ke password default berdasarkan
        role (e.g. Guru123! untuk guru, Siswa123! untuk siswa). Hanya dapat diakses
        oleh Admin.
      parameters:
      - description: ID User
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.SwaggerSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Reset password user ke password default
      tags:
      - User Management
  /api/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Mengaktifkan atau menonaktifkan user account berdasarkan ID. Hanya
        dapat diakses oleh Admin. Administrator tidak dapat dinonaktifkan.
      parameters:
      - description: ID User
        in: path
        name: id
        required: true
        type: integer
      - description: Payload status aktif
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.SwaggerSuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Perbarui status keaktifan user
      tags:
      - User Management
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"